Move iconName.ts and svgIcon.ts to common
