+++ packages/icons/src/common/svgIcon.ts
